refactor(TKT-328): consolidate ticket/tickets namespace to singular with --bulk flag
