+++ apps/cli/src/commands/ticket/spec.ts
@@ -4,13 +4,14 @@ import { PMOCommand, pmoBaseFlags, autoExportToBoard } from '../../lib/pmo/index
 import { styles } from '../../lib/styles.js';
 
 export default class TicketSpec extends PMOCommand {
-  static description = 'Assign a spec to a ticket';
+  static description = 'Assign a spec to ticket(s)';
 
   static examples = [
     '<%= config.bin %> <%= command.id %> TKT-001 SPEC-001',
     '<%= config.bin %> <%= command.id %> TKT-001',
     '<%= config.bin %> <%= command.id %>',
     '<%= config.bin %> <%= command.id %> TKT-001 --unlink',
+    '<%= config.bin %> <%= command.id %> --bulk --spec SPEC-001',
   ];
 
   static args = {
@@ -31,10 +32,25 @@ export default class TicketSpec extends PMOCommand {
       description: 'Remove spec from ticket instead of adding',
       default: false,
     }),
+    bulk: Flags.boolean({
+      char: 'b',
+      description: 'Enable bulk mode to assign spec to multiple tickets',
+      default: false,
+    }),
+    spec: Flags.string({
+      char: 's',
+      description: 'Spec ID to assign (for bulk mode)',
+    }),
   };
 
   async execute(): Promise<void> {
     const { args, flags } = await this.parse(TicketSpec);
+
+    // Bulk mode
+    if (flags.bulk) {
+      await this.executeBulk(flags);
+      return;
+    }
 
     // Get all tickets
     const tickets = await this.storage.listTickets();
@@ -168,5 +184,84 @@ export default class TicketSpec extends PMOCommand {
     this.log(styles.success(`\n✅ Linked ticket ${styles.emphasis(ticketId!)} to spec ${styles.emphasis(specId!)}`));
     this.log(styles.muted(`   Spec: ${spec.title}`));
     this.log(styles.muted(`\nView ticket: prlt ticket view ${ticketId}`));
+  }
+
+  private async executeBulk(flags: { spec?: string; unlink: boolean }): Promise<void> {
+    this.log(styles.emphasis('📄 Bulk Assign Spec to Tickets\n'));
+
+    // Get all tickets
+    const tickets = await this.storage.listTickets();
+    if (tickets.length === 0) {
+      this.log(styles.muted('\nNo tickets found.'));
+      return;
+    }
+
+    // Select tickets
+    const { selectedTickets } = await inquirer.prompt([{
+      type: 'checkbox',
+      name: 'selectedTickets',
+      message: 'Select tickets to update:',
+      choices: tickets.map(t => {
+        const specLabel = t.specId ? ` [spec: ${t.specId}]` : '';
+        return {
+          name: `${t.id} - ${t.title}${specLabel}`,
+          value: t.id,
+        };
+      }),
+    }]);
+
+    if (selectedTickets.length === 0) {
+      this.log(styles.muted('No tickets selected.'));
+      return;
+    }
+
+    // Handle unlink
+    if (flags.unlink) {
+      for (const ticketId of selectedTickets) {
+        await this.storage.updateTicket(ticketId, { specId: undefined });
+        this.log(styles.success(`  Unlinked spec from ${ticketId}`));
+      }
+      await autoExportToBoard(this.pmoPath, this.storage, (msg) => this.log(styles.muted(msg)));
+      this.log(styles.success(`\n✅ Unlinked spec from ${selectedTickets.length} ticket(s)`));
+      return;
+    }
+
+    // Get spec to assign
+    let specId = flags.spec;
+
+    if (!specId) {
+      const specs = await this.storage.listSpecs();
+      if (specs.length === 0) {
+        this.log(styles.muted('\nNo specs found. Create one with: prlt spec create'));
+        return;
+      }
+
+      const { selected } = await inquirer.prompt([{
+        type: 'list',
+        name: 'selected',
+        message: 'Select spec to assign:',
+        choices: specs.map(s => ({
+          name: `${s.id} - ${s.title} (${s.status})`,
+          value: s.id,
+        })),
+      }]);
+      specId = selected;
+    }
+
+    // Validate spec
+    const spec = await this.storage.getSpec(specId!);
+    if (!spec) {
+      this.error(`Spec not found: ${specId}`);
+    }
+
+    // Assign spec to all selected tickets
+    for (const ticketId of selectedTickets) {
+      await this.storage.updateTicket(ticketId, { specId });
+      this.log(styles.success(`  Linked ${ticketId} to ${specId}`));
+    }
+
+    await autoExportToBoard(this.pmoPath, this.storage, (msg) => this.log(styles.muted(msg)));
+
+    this.log(styles.success(`\n✅ Linked ${selectedTickets.length} ticket(s) to spec "${spec.title}"`));
   }
 }
